Implemented Short URL Route && Implemented ShortUrl Redirects
A second route is defined to catch any request that doesn't match the root
('/') url. This request is passed to the short url handler.

The short url handler tries to find a datastore entity that has the
supplied hash as a key. If it returns something it redircts otherwise
it returns a 404.

Fixes #10
Fixes #11