Added Some Design
In refernce to issue #13 the actual interface of the application needed a little work to make the application more presentable as well as define critical elements. With the addition of validation and error reporting it was important to highlight them with more severity than url results.

I added a static directory handler to the application configuration so that static css/images may now be served with the application. In addition a single main.css was created and is loaded with each page load.

It still looks hideous, maybe @robrbecker can spin some visual magic?